// src/app/shared/enum/response.enum.ts
export enum RESPONSE {
  SUCCESS = '2000',
  CREATED = '2001',
  BAD_REQUEST = '4000',
  TOO_MANY_PARAMETERS = '4003',
  INVALID_ACCESS_TOKEN = '4010',
  INVALID_PERMISSION = '4013',
  FAILED = '4011',
  UNAUTHORIZED = '4012',
  FORBIDDEN = '4030',
  NOT_FOUND = '4041',
  CONFLICT = '4090',
  INTERNAL_SERVER_ERROR = '5000',
  GATE_WAY_TIMEOUT = '5040',
  TOO_MANY_REQUESTS = '42900',
}

export enum ERR_CODE {
  REGISTRATION_REQUIRED = 'EB100',
  STORE_REQUIRED = 'EB101',
  USERNAME_ALREADY_EXISTS = 'EB120',
  EMAIL_ALREADY_EXISTS = 'EB121',
  NAME_ALREADY_EXISTS = 'EB122',
}